Use a -1 instead of int.max to prevent dropping a box pickup
There's a shortcut if the method is provided an invalid integer, basically negative values, that formally skip the box amount check entirely.  While having the player drop enough rounds to hit the int.max threshold is practically impossible, passing in a value that the method accounts for to skip the box altogether feels more proper.