Enhance Reusable Ammobox Spawning Logic
Centralized everything into the SpawnHandler, similar to Radtech's, HDBulletLib's, Hexadoken's, etc.
All Weapons that use a given ammo type will need to have their respective Reusable Ammobox updated here as well.
I'd love this to be automatic based on the rounds within the boxes or the originally spawned box but that seems impossible at the moment, so for now it's going to have to remain an opt-in configuration.