Add 4ga 00 Buckshot Box
It seems I've had this configured incorrectly the entire time so far, and there actually aren't any 4ga slug boxes, only 4ga 00 buckshot.  I've left the language entries and class definitions in case that fact ever changes.  This does mean I had to alter the 4mm box (which doesn't exactly exist...) to allow the 4ga shells to take the 4bx loadout code and take on the zbx since it's traditionally used for the ZM66.