Remove Redundant Usage definitions
Now that the SpawnHandler is doing this work, there's no need to define these relationships in the individual ammoboxes anymore.  In the future I'd like to expand on this "shared logic" and see what I can pull out of the SpawnHandler and instead define in the ReusableAmmobox Base Class, but for now this should suffice.