Enhance "Dropped Ammobox" logic instead
We might as well add a fallback here anyways.  If the item being replaced isn't a standard HDRoundAmmo it can't be bundled/boxed, so Rocquettes and other types of this ammo should simply attempt to be replaced like their map spawn logic would dictate, rather than trying to "remove" the box that they wouldn't realistically get without it being able to be boxed in the first place.